requirements.txt
flask
flask-cors
python-dotenv
cloudinary

fastapi
uvicorn

